recall_search_pattern
If the quickfix/location list has a stored search context, use it to
re-apply the corresponding search highlight.

This assumes that the "context" attribute has not been used for some
other purpose. Since it's a recent addition to the quickfix attributes,
it's hard to say exactly what purposes plugin authors will use it for,
but so far I haven't seen anyone using it at all, and @/ seems like the
most obvious use-case when the quickfix list is used to deal with search
results (as opposed to compilation errors).

It's worth noting that elements of the quickfix list have a "pattern"
attribute, but this lives on the individual list items, not the list as
a whole.