apply offset
Provide a way to use a different scrolloff setting when navigating with
the quickfix list than in other editing contexts. I generally like to
keep scrolloff at 1, but when jumping around via cnext/cprev it's nice
to keep a few more lines of context around the cursor. I'm exposing
listical#offset() to outside callers so it can be used in a <CR> map for
the qf filetype.