extract scroll-offset adjustment
I'm extracting the function that adjusts scroll offset to the coherent
plugin so it can be used in contexts other than listical. Listical will
broadcast an autocommand event that I can hook into to run the
adjustment when desired. This cleans up the listical s:go() code nicely
as well, since it no longer intertwines command and query.