conditional doautocmd
Unfortunately, Vim's doautocmd displays a "No matching autocommands"
message if nothing is listening for the event you're broadcasting. You
could suppress these by prefixing the command with "silent", but that
will also silence any messages that an event handler tries to display.

Another unfortunate thing about this is that this existence check
fails to recognize autocommands that are defined using certain pattern
mechanisms. For example, the following are two autocommand definitions
are equivalent,

    autocmd User FooBar,FooBaz call MyHandler()
    autocmd User Foo{Bar,Baz}  call MyHandler()

but exists('#User#FooBar') will return true for the first and false for
the second.