silent doautocmd
The problem with checking for autocommand existence is that certain
pattern constructs are not recognized. For example:

  autocmd User listical_next,listical_prev call foo
  autocmd User listical_{next,prev}        call foo

The first version is recognized by exists(), but the second version is
not. Since this plugin broadcasts multiple patterns, I'd rather risk
silencing a message that a hook tries to display than risk not
triggering the desired handler. Hooks generally aren't going to be
displaying messages anyway, and if they absolutely have to they can use
"unsilent" to do so.